backend/pyproject.toml
[project]
name = "mta_flask"
version = "0.0.3"
authors = [
  { name="Mark Colby", email="macolby14@gmail.com" },
]
description = "A python webserver that gives NYC MTA times"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
'Flask >= 2.2.2, <3',
'protobuf >= 4.21.12, <5',
'python-dotenv >= 0.21.0, <1',
'requests >= 2.28.1, <3',
'geopy >= 2.3.0, <3',
]

[project.optional-dependencies]
lint = [
  "flake8 >= 6.0.0, <7",
  "flake8-bugbear >= 23.2.3, <24",
  "black >= 22.12.0, <23",
  "isort >= 5.12.0, <6",
  "mypy >= 1.1.1, <2",
  "types-requests >= 2.28.0, <3",
]

[project.urls]
"Homepage" = "https://github.com/macolby14/mta-py"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
target-version = ['py39']

[tool.mypy]
exclude = [
    "mta_flask/proto/gtfs_realtime_pb2.py",
]
ignore_missing_imports = true
follow_imports = "silent"

[flake8]
exclude = [
  "mta_flask/proto/gtfs_realtime_pb2.py",
]
